% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{animate_system}
\alias{animate_system}
\title{Generate infection gif}
\usage{
animate_system(g_list, main, filepath)
}
\arguments{
\item{g_list}{A list of graphs representing each time step}

\item{main}{Title of the animated plot}

\item{filepath}{Name of the file}
}
\description{
Generate infection gif
}
