% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\name{createTimeline}
\alias{createTimeline}
\title{Process time increments and model the infection}
\usage{
createTimeline(g, t, model)
}
\arguments{
\item{g}{An \code{igraph}}

\item{t}{Number of time steps to compute}

\item{model}{An infection model}
}
\value{
A list of networks at each time step
}
\description{
Computes a time series of \code{t} steps on graph \code{g} as t_0 using model. The graph
must already be initialized. Run \code{initG(g, model)} prior to this function.
}
