% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityInfectionClass.R
\docType{class}
\name{denInfection-class}
\alias{denInfection-class}
\alias{density_infect}
\title{Density  Infection Class}
\description{
A default class to implement spreading an infection from initial nodes based on
infection density.
An \code{init_num} of nodes have their \code{$infected} property set to T.
When \code{$donext} the probability of infection is calculated using the sum of
adjacent infected nodes times the probability of infection.
}
\section{Fields}{

\describe{
\item{\code{init_num}}{Number of initial infected nodes}

\item{\code{transRate}}{Probability of infection given contact}
}}

\section{Methods}{

\describe{
\item{\code{donext(g)}}{Infect nodes adjacent to currently infected nodes}

\item{\code{init(g)}}{Initialize the graph with init_num infected nodes and set infected to red and susceptible to blue}
}}

