% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getStats}
\alias{getStats}
\title{Generate stats on SIR components}
\usage{
getStats(gCollection)
}
\arguments{
\item{gCollection}{List of networks at each timestep}
}
\description{
Computes the number of susceptible, infected, and recovered nodes in the graph
}
