% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_class.R
\docType{class}
\name{infection-class}
\alias{infection-class}
\alias{default_infect}
\title{Infection Class}
\description{
A default class to implement spreading an infection from initial nodes.
An \code{init_num} of nodes have their \code{$infected} property set to T.
When \code{$donext} is called each node adjacent to an infected node will
generate a random number. If this value is less than \code{rate}, the node
becomes infected.
}
\section{Fields}{

\describe{
\item{\code{init_num}}{Number of initial infected nodes}

\item{\code{rate}}{Probability an adjacent node will become infected}
}}

\section{Methods}{

\describe{
\item{\code{donext(g)}}{Infect nodes adjacent to currently infected nodes}

\item{\code{init(g)}}{Initialize the graph with init_num infected nodes and set infected to red and susceptible to blue}
}}

