% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_external.R
\docType{class}
\name{infection-external}
\alias{infection-external}
\alias{external_infect}
\title{Infection External Class}
\description{
An additional class to implement spreading an infection from external factors
or outside of the office. We initialize the graph with the existing nodes.
When donext is called a random number of nodes will be infected based
the external infection rate.
}
\section{Fields}{

\describe{
\item{\code{rate}}{External infection rate}
}}

\section{Methods}{

\describe{
\item{\code{donext(g)}}{Infect nodes randomly}

\item{\code{init(g)}}{Initialize all nodes with existing graph}
}}

