% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_model.R
\docType{class}
\name{infection_model-class}
\alias{infection_model-class}
\alias{infection_model}
\title{Infection Model}
\description{
An infection model collects the different components in the model and
executes them.
}
\section{Fields}{

\describe{
\item{\code{components}}{A list containing components of the model as reference classes}
}}

\section{Methods}{

\describe{
\item{\code{add_component(c)}}{Adds the component c to the list components}

\item{\code{init_model(g)}}{Initializes g using the init function of each component}

\item{\code{next_turn(g)}}{Modifies g using the donext function of each component}
}}

