% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\name{initG}
\alias{initG}
\title{Initialize graph with infected values}
\usage{
initG(g, model, onlyLarge = T)
}
\arguments{
\item{g}{An \code{igraph}}

\item{model}{An infection model initalized with components}

\item{onlyLarge}{Boolean indicating whether to prune \code{g} or not}
}
\value{
A graph \code{g} initialized with properties to run the model
}
\description{
Given a graph \code{g} and an infection model, prunes the disconnected components
from \code{g} and initializes the graph using the model components stored in model
}
