% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\name{keepLargeComponent}
\alias{keepLargeComponent}
\title{Isolate Largest Component}
\usage{
keepLargeComponent(g)
}
\arguments{
\item{g}{An \code{igraph}}
}
\value{
The graph \code{g} pruned of small disconnected components
}
\description{
Some networks generate small disconnected nodes or component. Use \code{keepLargeComponent} to
isolate the largest component so infection can spread in the model.
}
