% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\name{nextTurn}
\alias{nextTurn}
\title{Execute modification to next time step}
\usage{
nextTurn(g, model)
}
\arguments{
\item{g}{An \code{igraph}}

\item{model}{An infection model with components}
}
\value{
A graph \code{g} modified by actions stored in model
}
\description{
Run a single time step of the model on graph \code{g}
}
