% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery_class.R
\docType{class}
\name{recovery-class}
\alias{recovery-class}
\alias{default_recover}
\title{Default Recovery Class}
\description{
A default class to model a node's recovery. An infected node has a chance to
recover equal to counter/max_recovery_time. Counter keeps track the duration
of time a node has been infected. We set to -1 because donext automatically
increments counter and newly infected nodes this turn have a chance to
recover
}
\section{Fields}{

\describe{
\item{\code{max_recovery_time}}{Maximum time steps before a node is guaranteed to recover}
}}

\section{Methods}{

\describe{
\item{\code{donext(g)}}{Increments counter and computes whether a nodes state should be moved to recovered}

\item{\code{init(g)}}{Initialize all nodes with a counter and a recovered property}
}}

